+++ src/App.tsx
@@ -1,7 +1,5 @@
 import React, {useEffect, useState} from "react";
 import "./App.css";
-import KeysConfigForm from "./components/ConfigForm";
-import Home from "./components/Home";
 import {areAllConfigsValid} from "./util/ConfigValidator";
 import SpeechServiceConfiguration from "./models/SpeechServiceConfiguration";
 import QnAConfig from "./models/QnAConfig";
@@ -24,18 +22,19 @@ import {
 import TranslatorConfig from "./models/TranslatorConfig";
 import BingSearchConfig from "./models/BingSearchConfig";
 import {UseCaseModels} from "./util/UseCase";
-import UseCaseTemplate from "./components/usecases/UseCaseTemplate";
-import FourLangToSwissTranslation
-    from "./components/usecases/four_lang_to_swiss_translation/FourLangToSwissTranslation";
-import ChatWithBot from "./components/usecases/chat_with_bot/ChatWithBot";
-import RealtimeTranscription from "./components/usecases/realtime_transcription/RealtimeTranscription";
-import NewsReader from "./components/usecases/news_reader/NewsReader";
 import {ThemeProvider, createTheme} from "@mui/material/styles";
 import useTextToSpeech from "./hooks/useTextToSpeech";
 import {Voice} from "./util/TextToSpechVoices";
 import {Routes, Route} from "react-router-dom";
 import Layout from "./Layout";
 import {Alert, AlertTitle} from "@mui/material";
+import UseCaseTemplate from "./pages/usecases/UseCaseTemplate";
+import TranslationPage from "./pages/usecases/translation/TranslationPage";
+import QnaPage from "./pages/usecases/qna/QnaPage";
+import TranscriptionPage from "./pages/usecases/transcription/TranscriptionPage";
+import NewsPage from "./pages/usecases/news/NewsPage";
+import Home from "./pages/HomePage";
+import ConfigForm from "./pages/ConfigurationPage";
 
 export interface ICognitiveServicesConfiguration {
     speechServiceConfiguration: SpeechServiceConfiguration;
@@ -108,7 +107,7 @@ const App = () => {
                     <Route path="translate" element={
                         <UseCaseTemplate model={UseCaseModels.fourLangToSwissTranslation} error={useCaseError}
                                          synthesizeSpeech={synthesizeSpeech} isSynthesizing={isSynthesizing}>
-                            <FourLangToSwissTranslation
+                            <TranslationPage
                                 mySpeechConfig={cognitiveServicesConfiguration.speechServiceConfiguration}
                                 translatorConfig={cognitiveServicesConfiguration.translatorServiceConfiguration}
                                 setError={setUseCaseError}
@@ -119,7 +118,7 @@ const App = () => {
                     <Route path="chat" element={
                         <UseCaseTemplate model={UseCaseModels.botChat} error={useCaseError}
                                          synthesizeSpeech={synthesizeSpeech} isSynthesizing={isSynthesizing}>
-                            <ChatWithBot
+                            <QnaPage
                                 mySpeechConfig={cognitiveServicesConfiguration.speechServiceConfiguration}
                                 qnaConfig={cognitiveServicesConfiguration.qnaServiceConfiguration}
                                 translatorConfig={cognitiveServicesConfiguration.translatorServiceConfiguration}
@@ -131,7 +130,7 @@ const App = () => {
                     <Route path="transcription" element={
                         <UseCaseTemplate model={UseCaseModels.realtimeTranscription} error={useCaseError}
                                          synthesizeSpeech={synthesizeSpeech} isSynthesizing={isSynthesizing}>
-                            <RealtimeTranscription
+                            <TranscriptionPage
                                 speechConfig={cognitiveServicesConfiguration.speechServiceConfiguration}
                                 synthesizeSpeech={synthesizeSpeech}
                                 isSynthesizing={isSynthesizing}
@@ -141,13 +140,13 @@ const App = () => {
                     <Route path="newsreader" element={
                         <UseCaseTemplate model={UseCaseModels.newsReader} error={useCaseError}
                                          synthesizeSpeech={synthesizeSpeech} isSynthesizing={isSynthesizing}>
-                            <NewsReader
+                            <NewsPage
                                 synthesizeSpeech={synthesizeSpeech}
                                 isSynthesizing={isSynthesizing}
                                 setError={setUseCaseError}
                             />
                         </UseCaseTemplate>}/>
-                    <Route path="settings" element={<KeysConfigForm
+                    <Route path="settings" element={<ConfigForm
                         speechServiceConfig={cognitiveServicesConfiguration.speechServiceConfiguration}
                         qnaConfig={cognitiveServicesConfiguration.qnaServiceConfiguration}
                         translatorConfig={cognitiveServicesConfiguration.translatorServiceConfiguration}
