+++ src/pages/SpeechComponentProps.ts
