+++ src/pages/common/GravityItemsArea.tsx
